Signature-tree: 02d398910a4f865769d7e0215fad9bab9670faf8

+++ drivers/char/tpm/tpm.c
@@ -1035,7 +1035,7 @@ ssize_t tpm_write(struct file *file, const char __user *buf,
 
 	mutex_lock(&chip->buffer_mutex);
 
-	if (in_size > (unsigned int)TPM_BUFSIZE)
+	if (in_size > TPM_BUFSIZE)
 		in_size = TPM_BUFSIZE;
 
 	if (copy_from_user
